backend/entity/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Autenticazione e sicurezza
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Validazione dati
pydantic==2.5.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1

# Utilità
python-dotenv==1.0.0

# Testing (opzionale)
pytest==7.4.3
httpx==0.25.1
